# api/openapi.yaml
openapi: 3.0.3
info:
  title: Vehicle Analytics API Assessment
  version: 1.0.0
  description: >
    Only the health endpoint is predefined. You must design and implement the
    metadata and data API, then document your full API in this spec (or a separate
    file) and describe how to view it in justification.md.
paths:
  /health:
    get:
      tags:
        - Health
      operationId: getHealth
      summary: Health check (implemented in skeleton)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  emulator:
                    type: boolean
                example:
                  status: 'ok'
                  emulator: true
        '503':
          description: Unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Service is currently unhealthy, emulator could not be reached'
  /sensors:
    get:
      tags:
        - Sensors
      operationId: getSensors
      summary: Retrieve metadata of all sensors
      responses:
        '200':
          description: Successful retrieval of sensor metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorMetadata'
              example:
                - sensorId: 1000000070501
                  sensorName: 'BATTERY_TEMPERATURE'
                  unit: 'C'
                - sensorId: 2000000051201
                  sensorName: 'TYRE_PRESSURE_FL'
                  unit: 'kPa'
                - sensorId: 3000000060001
                  sensorName: 'PACK_CURRENT'
                  unit: 'A'
                - sensorId: 3000000060101
                  sensorName: 'VEHICLE_SPEED'
                  unit: 'km/h'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Unexpected server error'
        '503':
          description: No sensor metadata available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Sensor metadata not yet available'
  /sensors/{sensorId}:
    get:
      tags:
        - Sensors
      operationId: getSensor
      summary: Retrieve metadata of a sensor
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: number
          example: 1000000070501
          description: Unique identifier of the sensor
      responses:
        '200':
          description: Successful retrieval of sensor metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorMetadata'
              example:
                sensorId: 1000000070501
                sensorName: 'BATTERY_TEMPERATURE'
                unit: 'C'
        '400':
          description: Invalid sensor Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Invalid sensorId'
        '404':
          description: Sensor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Sensor not found'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Unexpected server error'
  /telemetry:
    get:
      tags:
        - Telemetry
      operationId: getLatestTelemetry
      summary: Retrieve latest telemetry data of all sensors
      responses:
        '200':
          description: Successful retrieval of telemetry data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorReading'
              example:
                - sensorId: 1000000070501
                  value: 33.7018539186185
                  timestamp: 1772577540.037
                - sensorId: 2000000051201
                  value: 190.96918605453672
                  timestamp: 1772577540.036
                - sensorId: 3000000060001
                  value: 421.56080856948626
                  timestamp: 1772577540.043
                - sensorId: 3000000060101
                  value: 17.8991355778208
                  timestamp: 1772577540.246
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Unexpected server error'
  /telemetry/stream:
    get:
      tags:
        - Telemetry
      operationId: streamTelemetry
      summary: Live SSE stream of telemetry readings
      description: >
        Opens a Server-Sent Events connection. The server pushes a 'reading'
        event for every telemetry message received from the emulator.
        Connect using the browser's native EventSource API.
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
              example: "event: reading\ndata: {\"sensorId\":1000000070501,\"value\":22.4,\"timestamp\":1234567890}\n\n"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Unexpected server error'
  /telemetry/{sensorId}:
    get:
      tags:
        - Telemetry
      operationId: getLatestSensorTelemetry
      summary: Retrieve latest telemetry data of a sensor
      parameters:
        - in: path
          name: sensorId
          required: true
          schema:
            type: number
          example: 1000000070501
          description: Unique identifier of the sensor
      responses:
        '200':
          description: Successful retrieval of telemetry data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReading'
              example:
                sensorId: 1000000070501
                value: 33.7018539186185
                timestamp: 1772577540.037
        '400':
          description: Invalid sensor Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Invalid sensorId'
        '404':
          description: No reading found for this sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Reading not found for sensorId'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Unexpected server error'
components:
  schemas:
    SensorMetadata:
      type: object
      properties:
        sensorId:
          type: number
        sensorName:
          type: string
        unit:
          type: string
      example:
        sensorId: 1000000070501
        sensorName: 'BATTERY_TEMPERATURE'
        unit: 'C'
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: 'An error occurred'
    SensorReading:
      type: object
      properties:
        sensorId:
          type: number
        value:
          type: number
        timestamp:
          type: number
      example:
        sensorId: 1000000070501
        value: 33.7018539186185
        timestamp: 1772577540.037
servers:
  - url: http://localhost:4000
    description: Local development server